+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Support orientation change for session replay (#4194)
 - Replay for crashes (#4171)
+- Redact web view from replay (#4203)
+- Add beforeCaptureViewHierarchy callback (#4210)
 
 ## 8.32.0